fix: compile errors on Godot 4.7-dev (double→float + obsolete dock API)
- CS0266 at SceneCollector.cs:492. AnimationPlayer.CurrentAnimationPosition
  returns a double in 4.7-dev (was float earlier). Store it as double
  — ap.Seek() takes a double already, so the rest of the bake loop
  works without further changes.

- CS0618 at PS1GodotPlugin.cs:50 and :83. AddControlToDock /
  RemoveControlFromDocks are marked obsolete in 4.7-dev in favor of
  AddDock(EditorDock) / RemoveDock(EditorDock). EditorDock is a brand-new
  class in dev and migrating would mean turning PS1GodotDock into an
  EditorDock subclass (currently VBoxContainer). The old API still
  works functionally — suppressing the warnings with #pragma
  warning disable CS0618 so warnings-as-errors builds pass. We'll
  migrate properly once 4.7 stabilizes or we pin to 4.4 (ROADMAP
  note in Phase 0.5 about stable-minor default).

+++ godot-ps1/addons/ps1godot/exporter/SceneCollector.cs
@@ -489,7 +489,7 @@ public static class SceneCollector
 
             // Stash the current play state so we restore it when done.
             var originalClip = ap.CurrentAnimation;
-            float originalPos = ap.CurrentAnimationPosition;
+            double originalPos = ap.CurrentAnimationPosition;
             ap.CurrentAnimation = clipName;
 
             for (int f = 0; f < frameCount; f++)
